HTMLtoPDf save dialouge box added

+++ SEGP/src/htmlToPdf/HTMLtoPDF.java
@@ -25,16 +25,20 @@ public class HTMLtoPDF {
 			// Show save file dialog
 			//The stage for show dialouge is get from MainClass stage
 			File file = fileChooser.showSaveDialog(MainClass.getStage());
+			
 
 			Client client = new Client("Hassan_Iqbal", "962ca71a838132f00947f2ec13984587");
+			if(file.isFile()){
 			FileOutputStream fileStream = new FileOutputStream(file);
-
+			
+			
 			String currentUrl = TabController.getWebEngine().getLocation().toString();
 			client.convertURI(currentUrl, fileStream);
 
 			fileStream.close();
-
+			}else{
 			System.out.println("File Download:check");
+			}
 			//
 		} catch (Exception e) {
 			e.printStackTrace();
